refactor: BindingAdapter 파일 통합
- PlaceListBindingAdapter.kt 파일을 삭제하고 해당 내용을 presentation/common/BindingAdapter.kt 로 이동
